requirements.txt
# Orchestration
apache-airflow==2.8.0

# Data processing
pandas==2.1.0
numpy==1.26.0
pydantic==2.5.0

# Cloud
google-cloud-storage==2.13.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# ML
scikit-learn==1.3.2
textblob==0.17.1
joblib==1.3.2
scipy==1.11.4

# API
fastapi==0.104.1
uvicorn==0.24.0

# Monitoring
streamlit==1.28.0

# HTTP
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
